backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9   # Uncomment if using PostgreSQL instead of SQLite
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
scikit-learn==1.5.2
numpy==1.26.4
python-multipart==0.0.9
